Fix problem with changing clothes, and losing stats, and also fixed porblem with leaving old unit alive
Refactory the C_change function to make sure old unit is disposed of
appropriately, also fix the bank initialization code so that the
bankaccount variable is tracked in stats for first time players.

The main issue with player losing their bank account and other stats
when changing clothes was that  when you changed clothes multiple times,
the variables list itself was not copied around to the new unit.