Add two extra options to admin menu
Add option in admin menu for setting an extra custom line in the
server's motd message.
Add option in the admin menu for deleting Man objects. This is useful
for when the clothes switching bugs out, and the old AI player stays.
Also reformatted the code in the admin menu to align the otions with
single tab.