Update masterarray.sqf
Changed the license check for cops on the civ vehicles that required the Crim Response because.... that was just dumb since they're not sold in cop base.

Removed the Sobr license check on regular an-2's for cops and replaced it with the air_support_training license. Did the opposite for the armed police an-2.